+++ Caf.Etl/Models/Manual/TidyData/DataTables/SoilGridPointSurveyV1.cs
@@ -10,8 +10,8 @@ namespace Caf.Etl.Models.Manual.TidyData.DataTables
         public int? ID2 { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
-        public int? TopDepth { get; set; }
-        public int? BottomDepth { get; set; }
+        public double? TopDepth { get; set; }
+        public double? BottomDepth { get; set; }
         public string Horizon { get; set; }
         public double? BulkDensity { get; set; }
         public double? dC13 { get; set; }
@@ -39,8 +39,8 @@ namespace Caf.Etl.Models.Manual.TidyData.DataTables
                    EqualityComparer<int?>.Default.Equals(ID2, other.ID2) &&
                    EqualityComparer<double?>.Default.Equals(Latitude, other.Latitude) &&
                    EqualityComparer<double?>.Default.Equals(Longitude, other.Longitude) &&
-                   EqualityComparer<int?>.Default.Equals(TopDepth, other.TopDepth) &&
-                   EqualityComparer<int?>.Default.Equals(BottomDepth, other.BottomDepth) &&
+                   EqualityComparer<double?>.Default.Equals(TopDepth, other.TopDepth) &&
+                   EqualityComparer<double?>.Default.Equals(BottomDepth, other.BottomDepth) &&
                    Horizon == other.Horizon &&
                    EqualityComparer<double?>.Default.Equals(BulkDensity, other.BulkDensity) &&
                    EqualityComparer<double?>.Default.Equals(dC13, other.dC13) &&
@@ -64,8 +64,8 @@ namespace Caf.Etl.Models.Manual.TidyData.DataTables
             hashCode = hashCode * -1521134295 + EqualityComparer<int?>.Default.GetHashCode(ID2);
             hashCode = hashCode * -1521134295 + EqualityComparer<double?>.Default.GetHashCode(Latitude);
             hashCode = hashCode * -1521134295 + EqualityComparer<double?>.Default.GetHashCode(Longitude);
-            hashCode = hashCode * -1521134295 + EqualityComparer<int?>.Default.GetHashCode(TopDepth);
-            hashCode = hashCode * -1521134295 + EqualityComparer<int?>.Default.GetHashCode(BottomDepth);
+            hashCode = hashCode * -1521134295 + EqualityComparer<double?>.Default.GetHashCode(TopDepth);
+            hashCode = hashCode * -1521134295 + EqualityComparer<double?>.Default.GetHashCode(BottomDepth);
             hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(Horizon);
             hashCode = hashCode * -1521134295 + EqualityComparer<double?>.Default.GetHashCode(BulkDensity);
             hashCode = hashCode * -1521134295 + EqualityComparer<double?>.Default.GetHashCode(dC13);
